+++ CHANGELOG.md
@@ -10,9 +10,12 @@ For the 2.x changelog see the [viur/server](https://github.com/viur-framework/se
 - Added validations to catch invalid recipient addresses early in sendEmail
 - 'connect-src': self and 'upgrade-insecure-requests' CSP directives by default
 - versionHash and appVersion variables to utils and jinja2 render 
+- The ability to import blobs that have been copied client-side from the old (non cloud-storage) blobstore
 
 ## Changed
+- [Breaking] srcSetFor function in jinja2 now needs a list with or height instead of deriving from groups
 - Replaced *.ggpht.com and *.googleusercontent.com CSP directives by storage.googleapis.com
+
 
 ### Fixed
 - AdminInfo for tree modules without a leaf skel